New timeslots on
Under "Show", "Talks on" was confusing since it sounded as if it would display only the talks occurring on that day of the week, whereas as actually it seems that it was the day that new timeslots would be shown as being on.  So I changed the text to "New timeslots on".